manage routes and relationships

+++ routes/web.php
@@ -18,37 +18,37 @@ use App\Http\Controllers\ListingController;
 // Get all listings
 Route::get('/', [ListingController::class, 'index']);
 
+// Show create listing form
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
+
 // Get single listing by id
 Route::get('/listings/{id}', [ListingController::class, 'show']);
 
-// Show create listing form
-Route::get('/listings/create', [ListingController::class, 'create']);
-
 // Store new listing data
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 // Show edit form
-Route::get('/listings/{id}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{id}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // Edit submit to update
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 // Delete listing
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
 // 
 
 // Show user registration form
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
 // Show login form
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 // Register new user
-Route::post('/users', [UserController::class, 'store']);
+Route::post('/users', [UserController::class, 'store'])->middleware('guest');
 
 // Login user
-Route::post('/users/authenticate', [UserController::class, 'authenticate']);
+Route::post('/users/authenticate', [UserController::class, 'authenticate'])->middleware('guest');
 
 // Logout
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
